+++ include/IrcServer.hpp
@@ -6,7 +6,7 @@
 /*   By: macarval <macarval@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/07 13:40:10 by macarval          #+#    #+#             */
-/*   Updated: 2024/06/27 15:23:15 by macarval         ###   ########.fr       */
+/*   Updated: 2024/06/27 18:07:14 by macarval         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,8 +30,8 @@
 #include "ChannelList.hpp"
 #include "FileTransfer.hpp"
 #include "Bot.hpp"
-#include "Colors.hpp"
-#include "numCode.hpp"
+#include "../utils/Colors.hpp"
+#include "../utils/numCode.hpp"
 
 class	Channel;
